% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_AOI.R
\docType{package}
\name{AOI-package}
\alias{AOI}
\alias{AOI-package}
\title{AOI: Areas of Interest}
\description{
An area of interest (AOI) is a geographic extent.
The aim of this package is to help users create these -
turning locations, place names, and political boundaries
into servicable representation for spatial analysis.
The package defaults to EPSG:4326\cr

See the \href{https://github.com/mikejohnson51/AOI}{README} on github,
and the project webpage for examples
\href{https://mikejohnson51.github.io/AOI/}{here}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mikejohnson51/AOI/}
  \item Report bugs at \url{https://github.com/mikejohnson51/AOI/issues}
}

}
\author{
\strong{Maintainer}: Mike Johnson \email{mikecp11@gmail.com} (\href{https://orcid.org/0000-0002-5288-8350}{ORCID})

Authors:
\itemize{
  \item Mike Johnson \email{mikecp11@gmail.com} (\href{https://orcid.org/0000-0002-5288-8350}{ORCID})
}

Other contributors:
\itemize{
  \item Justin Singh-Mohudpur \email{justin@justinsingh.me} (\href{https://orcid.org/0000-0002-5233-5799}{ORCID}) [contributor]
}

}
\keyword{internal}
