% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_AOI.R
\docType{package}
\name{AOI}
\alias{AOI}
\alias{AOI-package}
\title{AOI Package}
\description{
An area of interest (AOI) is a geographic extent. The aim of this package is to help users create these - essentailly turning locations and place names into servicable geometries.
The package is written using the simple features paradigm, however, by default, objects are returned as \code{SpatialPolygons} projected to EPSG:4269.
For those that have made the jump to sf, all functions include a 'sf' parameter that can be set to TRUE. Eventully the default behavior will change.\cr

The primary functions in this package are are \code{\link{geocode}}, \code{\link{revgeocode}}, \code{\link{getAOI}}, \code{\link{getBoundingBox}}.
The first returns a set spatial location from, place names using the OSM API; the second, a list of descritrpive featrues from a known loction;
the third retruns a single spatial geometry, and the last a geometry encompassing all input features.
Addtional helper functions include \code{\link{bbox_st}} and \code{\link{bbox_sp}} help convert AOIs between string and geometry represnetations;
\code{\link{check}} whcih helps users visualize AOIs in a interactive leaflet map; and \code{\link{modify}} allows for the modification of AOIs by uniform distances.
Finally, \code{\link{describe}} breaks existing spatial features into \code{\link{getAOI}} parameters to improve the reproducibility of geometry generation.
\cr

Two core datasets are served with the package. The first contains the spatial geometries and attributes of US \code{\link{states}} and the second contains the same for all US \code{\link{counties}}.
\cr

See the \href{https://github.com/mikejohnson51/AOI/blob/master/README.md}{README} on github, and a webpage of examples \href{https://mikejohnson51.github.io/AOI/}{here}.
}
