% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoi_describe.R
\name{aoi_describe}
\alias{aoi_describe}
\title{Describe an AOI}
\usage{
aoi_describe(AOI)
}
\arguments{
\item{AOI}{a spatial object (\code{raster}, \code{sf}, \code{sp}).}
}
\value{
a data.frame of AOI descriptors
}
\description{
Describe a spatial (sf/sp/raster) object in terms of a
             reproducible AOI  (e.g. \code{\link{aoi_get}}) parameters.
}
\examples{
{
  fname <- system.file("shape/nc.shp", package = "sf")
  nc <- sf::read_sf(fname)
  aoi_describe(AOI = nc[1, ])
}
}
