% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoi_draw.R
\name{aoi_draw}
\alias{aoi_draw}
\title{AOI Draw}
\usage{
aoi_draw()
}
\value{
An sf object called 'aoi'.
}
\description{
Interactivly draw an Area of Interest (AOI) using a shiny app. Once an object is drawn and the "Save AOI" button pressed,
a new sf object called 'aoi' will appear in you environment.
}
\examples{
\dontrun{
aoi_draw()
}


}
