% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoi_ext.R
\name{aoi_ext}
\alias{aoi_ext}
\title{AOI extent}
\usage{
aoi_ext(
  geo = NULL,
  xy = NULL,
  wh = NULL,
  units = default_units,
  crs = default_crs,
  bbox = FALSE
)
}
\arguments{
\item{geo}{an origion specificed by a name}

\item{xy}{a orign specified as a numeric vector}

\item{wh}{width and height (can be a single number) in units (see units arg)}

\item{units}{units of wh expansion}

\item{crs}{output crs}

\item{bbox}{return bbox object?}
}
\value{
vector or sf object
}
\description{
Build an extent surrinding by location point (longitude, latitude) based on a width and height.
}
