% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoi_get.R
\name{aoi_get}
\alias{aoi_get}
\title{Get Area of Interest (AOI) geometry}
\usage{
aoi_get(
  x = NULL,
  country = NULL,
  state = NULL,
  county = NULL,
  fip = NULL,
  zipcode = NULL,
  union = FALSE
)
}
\arguments{
\item{x}{\code{sf}, or a \code{Spat*} object}

\item{country}{\code{character}. Full name, ISO 3166-1 2 or 3 digit code.
Not case sensitive. Data comes from Natural Earth.}

\item{state}{\code{character}. Full name or two character abbreviation.
Not case sensitive. If \code{state = 'conus'}, the lower 48
states will be returned. If \code{state = 'all'}, all states
will be returned.}

\item{county}{\code{character}. County name(s). Requires \code{state} input.
Not case sensitive If 'all' then all counties in a state
are returned}

\item{fip}{a 2 or 5 digit US fip code}

\item{zipcode}{a US zip code. Will return a centroid.}

\item{union}{\code{logical}. If TRUE objects are unioned into a single object}
}
\value{
a sf geometry projected to \emph{EPSG:4326}.
}
\description{
Generate a spatial geometry from:
}
\examples{
\dontrun{
# Get AOI for a country
aoi_get(country = "Brazil")
# Get AOI defined by a state(s)
aoi_get(state = "CA")
aoi_get(state = c("CA", "nevada"))

# Get AOI defined by all states, or the lower 48
aoi_get(state = "all")
aoi_get(state = "conus")

# Get AOI defined by state & county pair(s)
aoi_get(state = "California", county = "Santa Barbara")
aoi_get(state = "CA", county = c("Santa Barbara", "ventura"))

# Get AOI defined by state & all counties
aoi_get(state = "California", county = "all")
}
}
