% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoi_utilities.R
\name{aoi_inside}
\alias{aoi_inside}
\title{Is Inside}
\usage{
aoi_inside(AOI, obj, total = TRUE)
}
\arguments{
\item{AOI}{object 2}

\item{obj}{object 1}

\item{total}{boolean. If \code{TRUE} then check if obj is competely inside the AOI (and vice versa: order doesn't matter)
If \code{FALSE}, then check if at least part of obj is in the AOI.}
}
\value{
boolean value
}
\description{
A check to see if one object is inside another
}
