% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoi_map.R
\name{aoi_map}
\alias{aoi_map}
\title{Generate Leafet map and tool set for AOI}
\usage{
aoi_map(AOI = NULL, returnMap = FALSE)
}
\arguments{
\item{AOI}{any spatial object (\code{raster}, \code{sf}, \code{sp}). Can be piped (\%>\%) from \code{\link{aoi_get}}. If \code{AOI = NULL}, base map of CONUS will be returned.}

\item{returnMap}{\code{logical}. If \code{FALSE} (default) the input AOI is returned and the leaflet map printed. If \code{TRUE} the leaflet map is returned and printed.}
}
\value{
a \code{leaflet} html object
}
\description{
Provides a precanned leaflet layout  for checking, and refining AOI queries. Useful \code{leaflet} tools allow for the marking of points, measuring of distances, and  panning and zooming.
}
\examples{
\dontrun{
## Generate an empty map:
     aoi_map()

## Check a defined AOI:
     AOI = getAOI(clip = list("UCSB", 10, 10))
     aoi_map(AOI)

## Chain to AOI calls:
     getAOI(clip = list("UCSB", 10, 10)) \%>\% aoi_map()

## Add layers with standard leaflet functions:
     r = getAOI("UCSB") \%>\%  # get AOI
         HydroData::findNWIS() # get SpatialPointsDataframe of local USGS gages

     aoi_map(r$AOI) \%>\%
        addMarkers(data = r$nwis, popup = r$nwis$site_no)

## Save map for reference:
     m = getAOI("Kansas City") \%>\% aoi_map()
     htmlwidgets::saveWidget(m, file = paste0(getwd(), "/myMap.html"))
}
}
