% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R, R/simpleCap.R, R/utility_aoiProj.R
\docType{data}
\name{aoiProj}
\alias{aoiProj}
\alias{aoiProj}
\alias{aoiProj}
\title{AOI Projection}
\format{An object of class \code{CRS} of length 1.}
\usage{
aoiProj

aoiProj

aoiProj
}
\description{
AOI Projection

AOI Projection

The projection used for all AOI calls \emph{EPSG:4269}
}
\author{
Mike Johnson

Mike Johnson

Mike Johnson
}
\keyword{datasets}
