% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbox_functions.R
\name{bbox_coords}
\alias{bbox_coords}
\title{Return bounding coordinates of a spatial (sp/sf/raster) object}
\usage{
bbox_coords(x)
}
\arguments{
\item{x}{a spatial object (sp/sf/raster)}
}
\value{
a data.frame
}
\description{
This function provides a wrapper around sf::st_bbox that instead returns a named
data.frame containing (xmin, ymin, xmax, ymax)
}
