% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbox_functions.R
\name{bbox_get}
\alias{bbox_get}
\title{Get Spatial Bounding Box}
\usage{
bbox_get(x)
}
\arguments{
\item{x}{input feature}
}
\value{
a sf polygon
}
\description{
Get spatial (sf) representation of bounding box of an input feature type.
Input can be data.frame, numeric, character, or spatial (sp/sf/raster).
If numeric or character order of inputs should be (xmin, xmax, ymin, ymax)
}
