% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbox_sp.R
\name{bbox_sp}
\alias{bbox_sp}
\title{Convert bounding box string to geometry}
\usage{
bbox_sp(bbox_st, sf = FALSE)
}
\arguments{
\item{bbox_st}{a bounding box string or vector in the order ("xmin","xmax", "ymin", "ymax")}

\item{sf}{logical. Should returned feature be of class sf (default = FALSE)}
}
\value{
a bounding box geometry
}
\description{
Convert a vector, dataframe, or bb object to a \code{SpatialPolygon}
}
\examples{
\dontrun{
CO = getAOI(state = 'CO') \%>\% bbox_st()
CO.1 = CO \%>\% bbox_sp()
}
}
\author{
Mike Johnson
}
