% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer.R
\name{buffer}
\alias{buffer}
\title{Buffer AOI}
\usage{
buffer(AOI, d, km = FALSE)
}
\arguments{
\item{AOI}{a spatial, raster or simple features object}

\item{d}{the distance by which to modify each edge}

\item{km}{is the distance in kilometers? Default is FALSE and in miles}
}
\value{
a spatial geometry of the same class as the input AOI (if Raster sp returned)
}
\description{
Add or subtract a uniform distance to/from a spatial obeject in either miles or kilometers.
}
\examples{
\dontrun{
 # get an AOI of 'Garden of the Gods' and add a 2 mile buffer
    getAOI("Garden of the Gods") \%>\% buffer(10)

 # get an AOI of 'Garden of the Gods' and add a 2 kilometer buffer
    getAOI("Garden of the Gods") \%>\% buffer(10, km = TRUE)

 # get and AOI for Colorado Springs and subtract 3 miles
    getAOI("Garden of the Gods") \%>\% buffer(-3)
}

}
\author{
Mike Johnson
}
