% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check}
\alias{check}
\title{Check AOI extents}
\usage{
check(AOI = NULL)
}
\arguments{
\item{AOI}{an AOI obtained from \code{getAOI}}
}
\value{
a leaflet map object
}
\description{
Check AOI extents
}
\author{
Mike Johnson
}
