% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_AOI.R
\docType{data}
\name{counties}
\alias{counties}
\title{USA Counites}
\format{a \code{SpatialPolygonsDataFrame}, 3220 observations of 7 variables
\itemize{
\item 'statefp':    A \code{character}  State 2-digit FederalInformationProcessingStandards (FIPS) code
\item 'countyfp':    A \code{character} County 3-digit FederalInformationProcessingStandards (FIPS) code
\item 'affgeoid': A \code{character}  AFF Summary Level Code
\item 'geoid':    A \code{character}  Concatinates state and county FIP code
\item 'name':     A \code{character}    County name
\item 'state_name':    A \code{character}    State name
\item 'state_abbr':     A \code{character}    State Abbriviation
}}
\source{
\href{https://cran.r-project.org/web/packages/USAboundaries/index.html}{USAboundaries}
}
\usage{
counties
}
\description{
Dataset containing \code{SpatialPolygons} of USA Counties. Data is initalized from the USAboundaries and USAboundariesData
package, converted to spatial \code{sp} objects, and cleaned-up for this package. The primary reason for doing this is to limit
the challenges associated with using the USAboundariesData (not on CRAN) as a dependency for this, and other packages, while also providing
a more minmualistic dataset.
}
\examples{
\dontrun{
 counties = AOI::counties
}
}
\keyword{datasets}
