% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_AOI.R
\docType{data}
\name{countries}
\alias{countries}
\title{Simplified World Boundaries}
\format{a \code{SpatialPolygonsDataFrame}, 246 observations of 11 variables:
\itemize{
\item 'FIPS'     :  \code{character} FIPS 10-4 Country Code
\item 'ISO2'     :  \code{character} ISO 3166-1 Alpha-2 Country Code
\item 'ISO3'     :  \code{character} ISO 3166-1 Alpha-3 Country Code
\item 'UN'       :  \code{integer}   ISO 3166-1 Numeric-3 Country Code
\item 'NAME'     :  \code{character} Name of country/area
\item 'AREA'     :  \code{integer}   Land area, FAO Statistics (2002)
\item 'POP2005'  :  \code{numeric}   Population, World Population Prospects (2005)
\item 'REGION'   :  \code{integer}   Macro geographical (continental region), UN Statistics
\item 'SUBREGION':  \code{integer}   Geographical sub-region, UN Statistics
\item 'LON'      :  \code{numeric}   Longitude
\item 'LAT'      :  \code{numeric}   Latitude
}}
\source{
\href{http://thematicmapping.org/downloads/world_borders.php}{thematicmapping.org}
}
\usage{
countries
}
\description{
Dataset containing simpified polygons of World Countries and Regions. Data is initalized from \href{http://thematicmapping.org/downloads/world_borders.php}{thematicmapping.org}.
}
\examples{
\dontrun{
 AOI::countries
}
}
\keyword{datasets}
