% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_AOI.R
\docType{package}
\name{default_crs}
\alias{default_crs}
\title{AOI Package}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
default_crs
}
\description{
An area of interest (AOI) is a geographic extent.
The aim of this package is to help users create these -
turning locations, place names, and political boundaries
into servicable representation for spatial analysis.
The package defaults to EPSG:4326\cr

See the \href{https://github.com/mikejohnson51/AOI}{README} on github,
and the project webpage for examples
\href{https://mikejohnson51.github.io/AOI/}{here}.
}
\keyword{datasets}
