% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_clip_unit.R
\name{define.clip.unit}
\alias{define.clip.unit}
\title{Interpret a clip unit from a user input}
\usage{
define.clip.unit(clip_unit)
}
\arguments{
\item{clip_unit}{a user supplied input}
}
\value{
a list of features defining the AOI
}
\description{
Interpret a clip unit from a user input
}
\author{
Mike Johnson
}
