% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{defineClip}
\alias{defineClip}
\title{definex}
\usage{
defineClip(x = NULL, km = FALSE)
}
\arguments{
\item{x}{a user supplied list (see \code{\link{aoi_get}})}

\item{km}{\code{logical}. If \code{TRUE} distance are in kilometers,
default is \code{FALSE} with distances in miles}
}
\value{
a 4-element list of features defining an AOI
}
\description{
Parse a x list from user input. \code{defineClip} parses user supplied lists to a format usable by \code{\link{getClip}}
}
\author{
Mike Johnson
}
\keyword{internal}
