% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definePoint.R
\name{definePoint}
\alias{definePoint}
\title{Define a Point}
\usage{
definePoint(point, geo = FALSE, coords = TRUE)
}
\arguments{
\item{point}{a point provided a lat,long vector or data.frame; or a string}

\item{geo}{\code{logical} return a `sf` point (Default `FALSE`)}

\item{coords}{\code{logical} return a `lat/long` data.frame (Default `TRUE`)}
}
\value{
a lat/lon data.frame and/or `sf` object
}
\description{
Return a lat/long data.frame and/or `sf` point from an input
}
\examples{
\dontrun{
definePoint("UCSB")
definePoint(c(37, -118))
definePoint("UCSB", geo = TRUE)
definePoint("UCSB", coords = FALSE)
}
}
\author{
Mike Johnson
}
