% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoi_ext.R
\name{discritize}
\alias{discritize}
\title{Materialize Grid from File or inputs}
\usage{
discritize(
  ext = NULL,
  dim = default_dim,
  in_crs = default_crs,
  out_crs = default_crs,
  spatrast = FALSE,
  fillValue = NULL,
  showWarnings = TRUE
)
}
\arguments{
\item{ext}{extent (xmin, xmax, ymin, ymax) in some coordinate system}

\item{dim}{dimension (number of columns, number of rows)}

\item{in_crs}{projection of input ext}

\item{out_crs}{projection of output object}

\item{spatrast}{should a SpatRaster object be returned? Default is FALSE}

\item{fillValue}{in spatrast is TRUE, what values should fill the object}

\item{showWarnings}{should warnings be shown?}
}
\value{
list or SpatRaster object
}
\description{
Materialize Grid from File or inputs
}
