% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoi_ext.R
\name{.domain}
\alias{.domain}
\title{Build Domain}
\usage{
.domain(xy, wh, units = default_units, crs = default_crs, bbox = FALSE)
}
\arguments{
\item{xy}{a orign specified as a numeric vector}

\item{wh}{width and height (can be a single number) in units (see units arg)}

\item{units}{units of wh expansion}

\item{crs}{output crs}

\item{bbox}{return bbox object?}
}
\value{
vector or sf object
}
\description{
Build Domain
}
