% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoCode.R
\name{.geocode}
\alias{.geocode}
\title{.geocode}
\usage{
.geocode(
  geo,
  pt = FALSE,
  bbox = FALSE,
  all = FALSE,
  method = default_method,
  crs = default_crs
)
}
\arguments{
\item{geo}{\code{character}. Place name(s)}

\item{pt}{\code{logical}. If TRUE point geometry is created.}

\item{bbox}{\code{logical}. If TRUE bounding box geometry is created}

\item{all}{\code{logical}. If TRUE the point, bbox and xy representations are returned as a list}

\item{method}{the geocoding service to be used. See ?tidygeocoder::geocode}

\item{crs}{desired CRS. Defaults to AOI::default_crs}
}
\value{
a data.frame, sf object, or vector
}
\description{
.geocode
}
