% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{fip_meta}
\alias{fip_meta}
\title{Returns a sf data.frame of fipio data}
\usage{
fip_meta(state, county = NULL)
}
\arguments{
\item{state}{State names, state abbreviations, or one of the following: "all", "conus", "territories"}

\item{county}{County names or "all"}
}
\value{
sf data.frame
}
\description{
Returns a sf data.frame of fipio data
}
\examples{
\dontrun{
fip_meta()
}
}
