% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_firstLower.R
\name{firstLower}
\alias{firstLower}
\title{Convert first character to lower case}
\usage{
firstLower(x)
}
\arguments{
\item{x}{a string to be parsed.}
}
\description{
\code{firstLower} lower cases the first character of a string
}
\examples{
\dontrun{
firstLower("SANTA BARBARA")
}
}
\author{
Mike Johnson
}
