% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoCode.R
\name{geocode_rev}
\alias{geocode_rev}
\title{Reverse Geocoding}
\usage{
geocode_rev(x, method = "osm")
}
\arguments{
\item{x}{a point provided by \code{numeric} lat/lon pair or
\code{character} place name}

\item{method}{"osm" (deafalt) or "ersi"}
}
\value{
a data.frame of descriptive features
}
\description{
Describe a location using the ERSI and OSM reverse geocoding web-services.
This service provides tradional reverse geocoding (lat/lon to placename)
but can also be use to get more information about a place name.
}
\examples{
\dontrun{
 geocode_rev(x = c(38,-115))
 geocode_rev("UCSB")
}
}
\author{
Mike Johnson
}
