% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode_wiki.R
\name{geocode_wiki}
\alias{geocode_wiki}
\title{Geocoding Events}
\usage{
geocode_wiki(event = NULL)
}
\arguments{
\item{event}{\code{character}. a term to search for on wikipeida}
}
\value{
aa data.frame of lat/lon coordinates
}
\description{
A wrapper around the Wikipedia API to return geo-coordinates of requested inputs.
}
\author{
Mike Johnson
}
