% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoCode.R
\name{geocodeAOI}
\alias{geocodeAOI}
\title{geocodeOSM}
\usage{
geocodeAOI(location, pt = FALSE, bb = FALSE, all = FALSE, method = "arcgis")
}
\arguments{
\item{location}{a place name}

\item{pt}{if TRUE a simple feature point is appended to returned list}

\item{bb}{if TRUE the OSM bounding area of the location is appended
to returned list}

\item{all}{if TRUE the point and bounding box representations are returned}

\item{method}{the geocoding service to be used. See ?tidygeocoder::geocode}
}
\value{
at minimum a data.frame of lat, long
}
\description{
Geocode via Open Street Maps API. \code{geocodeOSM}
takes an input string and converts it to a geolocation.
Additionally it can return the location as a simple
features point and the minimum bounding area
of the location extent.
}
\examples{
\dontrun{
geocodeAOI("UCSB")
geocodeAOI("Garden of the Gods", bb = TRUE)
}
}
