% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_geocodeGoogle.R
\name{geocodeGoogle}
\alias{geocodeGoogle}
\title{Geocode via Google Maps web-services}
\usage{
geocodeGoogle(location, pt = FALSE, bb = FALSE)
}
\arguments{
\item{location}{a place name}

\item{pt}{if TRUE a simple feature point is appended to returned list}

\item{bb}{if TRUE the OSM bounding area of the location is appended to returned list}
}
\value{
at minimum a data.frame of lat, long
}
\description{
\code{geocode} takes an input string and converts it to a geolocation.
Addtionally it can return the location as a simple features point and the minimun bounding area
of the location extent.
}
\examples{
\dontrun{
geocodeGoogle("UCSB")
geocodeGoogle("Garden of the Gods", bb = TRUE)
}
}
\author{
Mike Johnson
}
