% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFiat.R
\name{getFiat}
\alias{getFiat}
\title{Get County and State SpatialPolygon(s)}
\usage{
getFiat(state = NULL, county = NULL)
}
\arguments{
\item{state}{character. Full name or two character abbriviation. Not case senstive}

\item{county}{character. Provide county name(s). Requires 'state' input.}
}
\value{
a \code{SpatialPolygons} object projected to \emph{EPSG:4269}.
}
\description{
Get County and State SpatialPolygon(s)
}
\details{
\code{getFiat} returns a \code{SpatialPolygons} object for a defiend state and/or county.
Boundaries come from the 2017 US Census TIGER Dataset and are projected to \emph{EPSG:4269}.
}
\examples{
\dontrun{
# Get Single State
    getFiat(state = "CA")

# Get Multi-state
    getFiat(state = c("CA","Utah","Nevada"))

# Get County
    getFiat(state = "CA", county = "San Luis Obispo")

# Get Muli-county
   getFiat(state = "CA", county = c("San Luis Obispo", "Santa Barbara", "Ventura"))
}

}
\seealso{
\code{\link{getClip}}

\code{\link{getAOI}}
}
\author{
Mike Johnson
}
