% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPoint.R
\name{getPoint}
\alias{getPoint}
\title{Geocode a Point}
\usage{
getPoint(name = "UCSB")
}
\arguments{
\item{name}{a \code{character} place name}
}
\value{
a data.frame of lat,lon values
}
\description{
Overwrite the dismo::geocode function to only return lat/long, remove excessive print statements; and to retry on query errors.
}
\examples{
\dontrun{ getPoint("UCSB") }
}
\author{
Mike Johnson
}
