% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{data}
\name{HydroDataProj}
\alias{HydroDataProj}
\title{Variables needed for AOI calls}
\format{An object of class \code{CRS} of length 1.}
\usage{
HydroDataProj
}
\description{
Variables needed for AOI calls
}
\author{
Mike Johnson
}
\keyword{datasets}
