% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_inside.R
\name{is_inside}
\alias{is_inside}
\title{Is Inside}
\usage{
is_inside(AOI, obj, total = T)
}
\arguments{
\item{AOI}{object 2}

\item{obj}{object 1}

\item{total}{boolean. If \code{TRUE} then check if obj is competely inside the AOI.
If \code{FALSE}, then check if at least part of obj is in the AOI.}
}
\value{
boolean value
}
\description{
A check to see if one object is inside another
}
\examples{
\dontrun{
  AOI = getAOI(state = "CA")
  obj = getAOI(state = "CA", county = "Santa Barbara")
  is_inside(AOI, obj)

  AOI = getAOI(state = "CA")
  obj = getAOI(state = "CO", county = "El Paso")
  is_inside(AOI, obj)
}
}
\author{
Mike Johnson
}
