% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{make_polygon}
\alias{make_polygon}
\title{make polygons}
\usage{
make_polygon(north, east, south, west, crs = aoiProj)
}
\arguments{
\item{north}{max latitude}

\item{east}{max longitude}

\item{south}{min latitude}

\item{west}{min latitude}

\item{crs}{coordinate reference system (default aoiProj)}
}
\value{
sf polygon
}
\description{
turn coordinates into polygon
}
