% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoi_functions.R
\name{make_sf}
\alias{make_sf}
\title{Convert raster and sp objects to sf}
\usage{
make_sf(x)
}
\arguments{
\item{x}{any spatial object}
}
\value{
an sf object
}
\description{
Convert raster and sp objects to sf
}
\author{
Mike Johnson
}
\keyword{internal}
