% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nameAOI.R
\name{nameAOI}
\alias{nameAOI}
\title{Provide the plain-english description of an AOi}
\usage{
nameAOI(state = NULL, county = NULL, clip = NULL)
}
\arguments{
\item{state}{\code{character}.  Full name or two character abbriviation. Not case senstive}

\item{county}{\code{character}.  County name(s). Requires \code{state} input. Not case senstive}

\item{clip}{\code{Spatial} object, a \code{Raster} object, or a \code{list} (see details and \code{\link{getClip}})}
}
\value{
A character string
}
\description{
This function is a mirror of \code{\link{getAOI}} but instead of returning a \code{SpatilaPolygons} object,
returns a plain-english description of the AOI. (eg: 'AOI defined as a 10 mile tall by 10 mile wide region centered on (the) UCSB')
}
\examples{
\dontrun{
nameAOI(state = "CA")
}
}
\seealso{
\code{\link{nameClip}}
}
\author{
Mike Johnson
}
