% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nameClip.R
\name{nameClip}
\alias{nameClip}
\title{Interprete the plain-english description of a clip unit}
\usage{
nameClip(clip_unit)
}
\arguments{
\item{clip_unit}{a user supplied clip_unit}
}
\value{
a string describing the clipt unit in plain english
}
\description{
Interprete the plain-english description of a clip unit
}
\author{
Mike Johnson
}
