% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmbbSP.R
\name{osmbb_sp}
\alias{osmbb_sp}
\title{OSM bounding Box to Geometry}
\usage{
osmbb_sp(osm_bb)
}
\arguments{
\item{osm_bb}{an osm bounding box}
}
\value{
a \code{SpatialPolygon}
}
\description{
Convert a OSM returned bounding box to a \code{SpatialPolygon}
}
\author{
Mike Johnson
}
