% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_revGeo.R
\name{revGeo}
\alias{revGeo}
\title{Reverse Geocode}
\usage{
revGeo(point)
}
\arguments{
\item{point}{a point provided a lat,long vector or data.frame; or a string}
}
\value{
a data.frame of input point descriptions
}
\description{
Describe a point using the ERSI server geocoding  and reverse geocoding functionalities
}
\examples{
\dontrun{
pt1 = revGeo("UCSB")
print(pt1)

pt2 = revGeo(c(38,-115))
print(pt2)
}
}
\author{
Mike Johnson
}
