% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleCap.R
\name{simpleCap}
\alias{simpleCap}
\title{Camel Case}
\usage{
simpleCap(x)
}
\arguments{
\item{x}{a string to be parsed.}
}
\description{
\code{simpleCap} capitilizes all words in a given text string
}
\examples{
simpleCap("santa barbara")
}
\seealso{
Other AOI 'utility': \code{\link{firstLower}}
}
\author{
Mike Johnson
}
