% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{data}
\name{state.name}
\alias{state.name}
\title{Variables needed for AOI calls}
\format{An object of class \code{character} of length 50.}
\usage{
state.name
}
\description{
Variables needed for AOI calls
}
\author{
Mike Johnson
}
\keyword{datasets}
