% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{data}
\name{stateAbb}
\alias{stateAbb}
\title{State Abbrivations}
\format{An object of class \code{character} of length 50.}
\usage{
stateAbb
}
\description{
State Abbrivations
}
\author{
Mike Johnson
}
\keyword{datasets}
