% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{data}
\name{stateName}
\alias{stateName}
\title{State Names}
\format{An object of class \code{character} of length 50.}
\usage{
stateName
}
\description{
State Names
}
\author{
Mike Johnson
}
\keyword{datasets}
