% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_AOI.R
\docType{data}
\name{zipcodes}
\alias{zipcodes}
\title{USA Zipcode Database}
\format{a \code{csv}, 246 43191 of 7 variables:
\itemize{
\item 'zip'     :  \code{integer} Zipcode
\item 'city'     :  \code{character} City Name
\item 'state'     :  \code{character} State Name
\item 'latitude'       :  \code{numeric}   latitude
\item 'longitude'     :  \code{numeric} longitude
\item 'timezone'     :  \code{integer}  timezone (from UTC)
\item 'dst'  :  \code{integer}   Daylights saving flag
}}
\source{
\href{https://www.boutell.com/zipcodes/}{Tom Boutell}
}
\usage{
zipcodes
}
\description{
Dataset containing information for USA Zipcodes. Data is initalized from \href{https://www.boutell.com/zipcodes/}{Tom Boutell}
}
\examples{
\dontrun{
 AOI::zipcodes
}
}
\keyword{datasets}
