% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode_wiki.R
\name{alt_page}
\alias{alt_page}
\title{Alternate Page Finder}
\usage{
alt_page(loc, pt = FALSE)
}
\arguments{
\item{loc}{a wikipedia structured call}

\item{pts}{\code{logical}. If TRUE point geometery is appended to the returned list()}
}
\value{
at minimum a data.frame of lat, long
}
\description{
Find linked pages to a wikipedia call
}
\examples{
\dontrun{
alt_page("Twin_towers")
}
}
\author{
Mike Johnson
}
\keyword{internal}
