% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_AOI.R
\docType{package}
\name{AOI}
\alias{AOI}
\alias{AOI-package}
\title{AOI}
\description{
\code{AOI} package
The area of interest (AOI) is the geographic extent. It helps confin the unit of work to a geographic area,
tproritize and define research and subsetting efforts, and to improve reproducabilty across studies.
This package aims to make finding fiat and geographic AOIs easier,
through a common query system based on 'state', 'county' and 'clip' parameters.
AOIs for all queryies are retruned a a SpatialPolygon projeted to EPSG:4269.

See the \href{https://github.com/mikejohnson51/AOI/blob/master/README.md}{README} on github
}
