% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AOI.R
\docType{package}
\name{AOI}
\alias{AOI}
\alias{AOI-package}
\title{AOI}
\description{
\code{AOI} package
}
\details{
The area of interest (AOI) is the geographic extent of a project.
This helps confine the unit of work to a geographic area, and helps to not only proritize and define research and subsetting efforts,
but to improve reproducabilty across studies. This package aims to make finding state, county and geographic AOI easier,
through a common query system based on 'state', 'county' and 'clip' parameters. AOIs for all queryies are retruned a a SpatialPolygon.

See the README on github
}
