% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_AOI.R
\docType{package}
\name{AOI}
\alias{AOI}
\title{AOI Package}
\description{
An area of interest (AOI) is a geographic extent. The aim of this package is to help users create these -  turning locations, place names, and political boundaries
into servicable geometries for spatial analysis. The package is written using the simple features paradigm, projected to EPSG:4269.\cr

The primary functions are \code{\link{geocode}}, \code{\link{geocode_rev}}, \code{\link{aoi_get}}, and \code{\link{bbox_get}}.
The first returns a data.frame of corrdinates from place names using the OSM API; the second returns a list of descriptive features from a known place name or lat/lon pair;
the third returns a spatial (sf) geometry from a country, state, county, or defined region, and the last an extent encompassing a set of input features.
\code{\link{aoi_map}} helps users visualize AOIs in a interactive map; and \code{\link{aoi_buffer}} which allows AOIs to be modified by uniform distances.
Finally, \code{\link{aoi_describe}} breaks existing spatial features into \code{\link{aoi_get}} parameters to improve the reproducibility of geometry generation.
\cr

See the \href{https://github.com/mikejohnson51/AOI}{README} on github, and the project webpage for examples \href{https://mikejohnson51.github.io/AOI/}{here}.
}
