% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoi_describe.R
\name{aoi_describe}
\alias{aoi_describe}
\title{Describe an AOI}
\usage{
aoi_describe(AOI, full = FALSE, km = FALSE)
}
\arguments{
\item{AOI}{a spatial object (\code{raster}, \code{sf}, \code{sp}).}

\item{full}{if TRUE, reverse geocoding descriptions returned, else just lat, lon, width, height, and origin (default = FALSE)}

\item{km}{if TRUE, units are in kilometers, else in miles (default = FALSE)}
}
\value{
a data.frame of AOI descriptors including (at minimum):
\describe{
  \item{lat}{the AOI center latitude }
  \item{lon}{the AOI center longitude}
  \item{height}{ height in (miles)}
  \item{width}{width in(miles)}
  \item{origin}{AOI origin}
}
}
\description{
Describe a spatial (sf/sp/raster) object in terms of a reproducable AOI  (e.g. \code{\link{aoi_get}}) parameters.
}
\examples{
 {library(AOI)
 aoi_get("UCSB") \%>\% aoi_describe()
 aoi_get("UCSB") \%>\% aoi_describe(full = TRUE)
 }
}
