% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_aoiProj.R
\docType{data}
\name{aoiProj}
\alias{aoiProj}
\title{AOI Projection}
\format{An object of class \code{character} of length 1.}
\usage{
aoiProj
}
\description{
Base projection used for all AOI calls: \emph{EPSG:4269}. `aoiProj = "+init=epsg:4269"`
}
\author{
Mike Johnson
}
\keyword{datasets}
