% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{data}
\name{aoiProj}
\alias{aoiProj}
\title{AOI Projection}
\format{An object of class \code{CRS} of length 1.}
\usage{
aoiProj
}
\description{
AOI Projection
}
\author{
Mike Johnson
}
\keyword{datasets}
