% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_aoiProj.R
\docType{data}
\name{aoiProj}
\alias{aoiProj}
\title{AOI Projection}
\format{An object of class \code{CRS} of length 1.}
\usage{
aoiProj
}
\description{
The projection used for all AOI calls: \emph{EPSG:4269}
}
\author{
Mike Johnson
}
\keyword{datasets}
