% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_conversions.R
\name{bbox_sp}
\alias{bbox_sp}
\title{Convert bounding box strings to spatail geometries}
\usage{
bbox_sp(s)
}
\arguments{
\item{s}{a comma seperated \code{character}  string, \code{numeric} vector, or data.frame in the order ("xmin","xmax", "ymin", "ymax"). \code{Raster} objects are also accepted.}
}
\value{
a spatial (sp/sf) geometry projected to \emph{EPSG:4269}
}
\description{
Convert a vector, data.frame, bb object, or raster to a spatial (sp/sf) geometry
}
\examples{
\dontrun{

## SpatialPolygon from string
   bbox = bbox_sp("37,36,-119,-118")

## SpatialPolygon from vector
   bbox = c(37,38,-119,-118) \%>\% bbox_sp()

## Simple Feature Polygon from data.frame
   bbox = data.frame(xmin = 37, xmax = 38, ymin = -119, ymax = -118) \%>\% bbox_sp(sf = T)

## SpatialPolygon from Reverse Geocoding results
   bbox = revgeocode("Santa Barbara")$bb \%>\% bbox_sp()

## String to Geometry to String (full circle)
    bbox = c(37,38,-119,-118) \%>\% bbox_sp() \%>\% bbox_st()

## Raster to simple features polygon
   raster \%>\% bbox_sp(sf = TRUE)
}
}
\seealso{
\code{\link{bbox_st}}
}
\author{
Mike Johnson
}
