% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbox_sp.R
\name{bbox_sp}
\alias{bbox_sp}
\title{Convert bounding box string to geometry}
\usage{
bbox_sp(bbox_st, sf = FALSE)
}
\arguments{
\item{bbox_st}{a comma seperated \code{character}  string, \code{numeric} vector, or data.frame in the order ("xmin","xmax", "ymin", "ymax"). \code{Raster} objects are also accepted.}

\item{sf}{\code{logical}. If TRUE object returned is of class \code{sf}, default is FALSE and returns \code{SpatialPolygons}}
}
\value{
a bounding box geometry
}
\description{
Convert a vector, dataframe, bb, or raster object to a spatial geometry
}
\examples{
\dontrun{

## SpatialPolygon from string
   bbox = bbox_sp("37,36,-119,-118")

## SpatialPolygon from vector
   bbox = c(37,38,-119,-118) \%>\% bbox_sp()

## Simple Feature Polygon from data.frame
   bbox = data.frame(xmin = 37, xmax = 38, ymin = -119, ymax = -118) \%>\% bbox_sp(sf = T)

## SpatialPolygon from Reverse Geocoding results
   bbox = revgeocode("Santa Barbara")$bb \%>\% bbox_sp()

## String to Geometry to String (full circle)
    bbox = c(37,38,-119,-118) \%>\% bbox_sp() \%>\% bbox_st()

## Raster to sf
   raster \%>\% bbox_sp(sf = TRUE)
}
}
\author{
Mike Johnson
}
