% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbox_st.R
\name{bbox_st}
\alias{bbox_st}
\title{Convert bounding box geometry to string}
\usage{
bbox_st(AOI)
}
\arguments{
\item{AOI}{an AOI obtained using \link{getAOI} (or any sp/sf object).}
}
\value{
a bounding box data.frame
}
\description{
Convert an AOI or spatial object to a data.frame of (xmin, xmax, ymin, ymax)
}
\examples{
\dontrun{
#Get a bounding box data.frame for AOI
   AOI = getAOI(clip = list("UCSB", 10, 10))
   bb = bbox_st(AOI)
print(bb)

# Chain to AOI calls:
AOI = getAOI(clip = list("UCSB", 10, 10)) \%>\% bbox_st()
print(AOI)
}
}
\author{
Mike Johnson
}
