% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbox_st.R
\name{bbox_st}
\alias{bbox_st}
\title{Convert spatial geometry to a data.frame}
\usage{
bbox_st(AOI)
}
\arguments{
\item{AOI}{any spatial object (\code{raster}, \code{sf}, \code{sp}). Can be piped (\%>\%) from \code{\link{getAOI}}}
}
\value{
a data.frame containing xmin. xmax, ymin, ymax coordinates
}
\description{
Convert a spatial object to a data.frame of (xmin, xmax, ymin, ymax)
}
\examples{
\dontrun{
## Get a bounding box data.frame for AOI
   AOI = getAOI(list("UCSB", 10, 10)) \%>\% bbox_st()

>   xmin     xmax    ymin     ymax
>  -119.9337 -119.758 34.34213 34.48706
}
}
\author{
Mike Johnson
}
