% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check}
\alias{check}
\title{Vizualize AOIs}
\usage{
check(AOI = NULL)
}
\arguments{
\item{AOI}{an AOI obtained using \link{getAOI} or any/sf. Can be left \code{NULL}}
}
\value{
a list of AOI and \code{leaflet} html object
}
\description{
Generate an interactive \code{leaflet} map for defining, checking, and refining AOI queries. Can be chained to \link{getAOI} via
`%>%`. Useful \code{leaflet} tools allow for the marking of points, measuring of distances, and interactive panning and zooming to help define
an approapriate AOI.
}
\examples{
\dontrun{
# Generate an empty map:
check()

# Check a defined AOI:
AOI = getAOI(clip = list("UCSB", 10, 10))
check(AOI)

# Chain to AOI calls:
getAOI(clip = list("UCSB", 10, 10)) \%>\% check()
}
}
\author{
Mike Johnson
}
