% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_check_class.R
\name{checkClass}
\alias{checkClass}
\title{Check an objects class}
\usage{
checkClass(x, type)
}
\arguments{
\item{x}{an object}

\item{type}{a class}
}
\value{
logical
}
\description{
A function to check the class of an object, will return TRUE if `x`` is of class `type``
}
\examples{
sf = getAOI(state = "CA", sf = TRUE)
checkClass(sf, "sf")
}
\author{
Mike Johnson
}
