% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_check_class.R
\name{checkClass}
\alias{checkClass}
\title{Check Objects Class}
\usage{
checkClass(x, type)
}
\arguments{
\item{x}{an object}

\item{type}{a class}
}
\value{
logical
}
\description{
A function check the class of an object, will return TRUE if x is of class type
}
\examples{
\dontrun{
sf = getAOI(state = "CA", sf = TRUE)
checkClass(sf, "sf")
}

}
\seealso{
Other HydroData 'utility' function: \code{\link{getBoundingBox}}
}
\author{
Mike Johnson
}
