% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_AOI.R
\docType{data}
\name{counties}
\alias{counties}
\title{USA Counites}
\format{a \code{SpatialPolygonsDataFrame}, 3220 observations of 7 variables:
\itemize{
\item 'statefp'   : \code{character} 2-digit Federal Information Processing Standards (FIPS) code
\item 'countyfp'  : \code{character} 3-digit Federal Information Processing Standards (FIPS) code
\item 'affgeoid'  : \code{character} American FactFinder (AFF) Summary Level Code
\item 'geoid'     : \code{character} Concatinates state and county FIP code
\item 'name'      : \code{character} County name
\item 'state_name': \code{character} State name
\item 'state_abbr': \code{character} State Abbriviation
}}
\usage{
counties
}
\description{
Dataset containing polygon representations of USA Counties.
}
\examples{
\dontrun{
 AOI::counties
}
}
\keyword{datasets}
