% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineClip.R
\name{defineClip}
\alias{defineClip}
\title{Parse a clip list from user input}
\usage{
defineClip(clip = NULL, km = FALSE)
}
\arguments{
\item{clip}{a user supplied list (see \code{\link{getAOI}})}

\item{km}{\code{logical}. If \code{TRUE} distance are in kilometers,
default is \code{FALSE} with distances in miles}
}
\value{
a 4-element list of features defining an AOI
}
\description{
\code{defineClip} parses user supplied lists to a format usable by \code{\link{getClip}}
}
\author{
Mike Johnson
}
