% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineClip.R
\name{defineClip}
\alias{defineClip}
\title{Parse a clip list from user input}
\usage{
defineClip(clip = NULL)
}
\arguments{
\item{clip}{a user supplied list (see \code{\link{getAOI}})}
}
\value{
a 4-element list of features defining an AOI including:
}
\description{
\code{defineClip} parses user supplied lists to a format usable by \code{\link{getClip}}
}
\examples{
\dontrun{
defineClip(list("UCSB", 10, 10))
defineClip(list(37,-115, 10, 10))
}

}
\seealso{
\code{\link{getClip}}

\code{\link{getAOI}}
}
\author{
Mike Johnson
}
