% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe}
\alias{describe}
\title{Describe an AOI}
\usage{
describe(AOI, full = FALSE)
}
\arguments{
\item{AOI}{any spatial object (\code{raster}, \code{sf}, \code{sp}). Can be piped (\%>\%) from \code{\link{getAOI}}}

\item{full}{if TRUE reverse geocoding descriptions returned, else just location, width, height, and origin (defualt = FALSE)}
}
\value{
a data.frame of AOI descriptors including:
\describe{
  \item{latCent}{the AOI center latitude }
  \item{lngCent}{the AOI center longitude}
  \item{height}{ height in (miles)}
  \item{width}{width in(miles)}
  \item{origin}{AOI origin}
  \item{name}{Most descriptive place name from \code{\link{revgeocode}}}
}
}
\description{
Describes a spatial, raster or sf object in terms of a reproducable clip area.
}
\examples{
\dontrun{
 AOI = getAOI(clip = list("UCSB", 10, 10)) \%>\% describe()

 ```
 latCent :	34.4139629
 lngCent :	-119.848947
 height  :	10 miles
 width   :	10 miles
 origin  :	center
 name    :	93106, Santa Barbara, California
 ```
}
}
\author{
Mike Johnson
}
