% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe}
\alias{describe}
\title{Describe an AOI}
\usage{
describe(AOI)
}
\arguments{
\item{AOI}{an AOI obtained using \link{getAOI}.}
}
\value{
a data.frame of AOI descriptors including
\describe{
  \item{latCent}{the AOI center latitude}
  \item{lngCent}{the AOI center longitude}
  \item{height}{ height in (miles)}
  \item{width}{width in(miles)}
  \item{origin}{AOI origin}
  \item{name}{Most descriptive geocoded name from \code{geoCode}}
}
}
\description{
Convert an AOI object to a data.frame of describing factors. Can be usefull for sharing,
documenting and repaeating AOI calls.
}
\examples{
\dontrun{
#Get an AOI
AOI = getAOI(clip = list("UCSB", 10, 10))
describe(AOI)

# Chain to AOI calls:
AOI = getAOI(clip = list("UCSB", 10, 10)) \%>\% describe()
}
}
\author{
Mike Johnson
}
