% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe}
\alias{describe}
\title{Describe an AOI}
\usage{
describe(AOI, full = FALSE)
}
\arguments{
\item{AOI}{any spatial object (\code{raster}, \code{sf}, \code{sp}).}

\item{full}{if TRUE, reverse geocoding descriptions returned, else just lat, lon, width, height, and origin (default = FALSE)}
}
\value{
a data.frame of AOI descriptors including (at minimum):
\describe{
  \item{lat}{the AOI center latitude }
  \item{lon}{the AOI center longitude}
  \item{height}{ height in (miles)}
  \item{width}{width in(miles)}
  \item{origin}{AOI origin}
}
}
\description{
Describe a spatial, raster or sf object in terms of a reproducable clip area (e.g. \code{\link{getAOI}}) parmaters.
}
\examples{
\dontrun{
 AOI = getAOI("UCSB") \%>\% describe()

 ```
 lat     :	34.41456
 lon     :	-119.8605796
 height  :	1
 width   :	3
 origin  :	center
 ```

 AOI = getAOI("UCSB") \%>\% describe(full = TRUE)

 ```
 lat     :	34.41456
 lon     :	-119.8605796
 height  :	1
 width   :	3
 origin  :	center
 name    :	6650 Abrego Rd, Goleta, California, 93117
 area    :	3 square miles
 ```
}
}
\author{
Mike Johnson
}
