% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe}
\alias{describe}
\title{Describe an AOI}
\usage{
describe(AOI)
}
\arguments{
\item{AOI}{a spatial, raster or sf object}
}
\value{
a data.frame of AOI descriptors including
\describe{
  \item{latCent}{the AOI center latitude}
  \item{lngCent}{the AOI center longitude}
  \item{height}{ height in (miles)}
  \item{width}{width in(miles)}
  \item{origin}{AOI origin}
  \item{name}{Most descriptive geocoded name from \code{\link{revgeocode}}}
}
}
\description{
Breaks a Spatial object into the features describing a reporducable clip area.
}
\examples{
\dontrun{
 AOI = getAOI(clip = list("UCSB", 10, 10)) \%>\% describe()

 ```
 latCent :	34.4139629
 lngCent :	-119.848947
 height  :	10 miles
 width   :	10 miles
 origin  :	center
 name    :	93106, Santa Barbara, California
 ```

}
}
\author{
Mike Johnson
}
