% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleCap.R, R/utility_firstLower.R
\name{firstLower}
\alias{firstLower}
\alias{firstLower}
\title{Lower case first character}
\usage{
firstLower(x)

firstLower(x)
}
\arguments{
\item{x}{a string to be parsed.}

\item{x}{a string to be parsed.}
}
\description{
\code{simpleCap} lower cases the first character of a string

\code{firstLower} lower cases the first character of a string
}
\examples{
simpleCap("santa barbara")
firstLower("SANTA BARBARA")
}
\seealso{
Other AOI 'utility': \code{\link{simpleCap}}

Other AOI 'utility': \code{\link{simpleCap}}
}
\author{
Mike Johnson

Mike Johnson
}
