% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleCap.R
\name{firstLower}
\alias{firstLower}
\title{Lower case first character}
\usage{
firstLower(x)
}
\arguments{
\item{x}{a string to be parsed.}
}
\description{
\code{simpleCap} lower cases the first character of a string
}
\examples{
simpleCap("santa barbara")
}
\seealso{
Other AOI 'utility': \code{\link{simpleCap}}
}
\author{
Mike Johnson
}
