% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstLower.R, R/utility_firstLower.R
\name{firstLower}
\alias{firstLower}
\alias{firstLower}
\title{Lower case first character}
\usage{
firstLower(x)

firstLower(x)
}
\arguments{
\item{x}{a string to be parsed.}

\item{s}{a string to be parsed.}
}
\description{
\code{simpleCap} lower cases the first character of a string

\code{firstLower} lower cases the first character of a string
}
\examples{
/dontrun{ simpleCap("santa barbara") }
firstLower("SANTA BARBARA")
}
\author{
Mike Johnson

Mike Johnson
}
