% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{geocode}
\alias{geocode}
\title{Geocoding}
\usage{
geocode(location = NULL, pt = FALSE, bb = FALSE, sf = FALSE)
}
\arguments{
\item{location}{\code{character}. Place name(s)}

\item{pt}{\code{logical}. If TRUE point geometery is appended to the returned list()}

\item{bb}{\code{logical}. If TRUE bounding box geometry is appended to the returned list()}

\item{sf}{\code{logical}. If TRUE object(s) returned are of class \code{sf}, default is FALSE and returns \code{sp}}
}
\value{
at minimum a matrix of lat/lon coordinates. Possible list with appended spatial features of type \code{sf} or \code{sp}
}
\description{
A wrapper around the OpenSteetMap geocoding web-services. Users can request a lat/lon pair, spatial points, and/or a bounding box geometries.\cr\cr
One or more locations can be given at a time. If a single point is requested, `geocode` will provide a matix of lat/lon values; a spatial point and the geocode derived bounding box (if requested).
If multiple points are given the returned objects will be a matrix with columns for input name-lat-lon; a SpatialPoints object; and a minimum bounding box of all input locations.
}
\examples{
\dontrun{
 ## geocode a single location
     geocode("UCSB")
 ## geocode a single location and return a SpatialPoints object
     geocode("UCSB", pt = TRUE)
 ## geocode a single location and derived bounding box of location
     geocode("UCSB", bb = TRUE)
 ## geocode multiple locations
     geocode(c("UCSB", "Goleta", "Sterns Warf"))
 ## geocode multiple points and generate a minimum bounding box of all locations
     geocode(c("UCSB", "Goleta", "Sterns Warf"), bb = T, pt= T)
}
}
\author{
Mike Johnson
}
