% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{geocode}
\alias{geocode}
\title{Geocoding}
\usage{
geocode(location = NULL, pt = FALSE, bb = FALSE, server = "google")
}
\arguments{
\item{location}{place name(s)}

\item{pt}{logical. Should the function return a spatial feature(s) of the location}

\item{bb}{return bb Should a bounding box geometery be returned with the object.}

\item{server}{what server should be prioritized. Options inlcude "google" or "OSM" (default = 'google)}
}
\value{
at minimum a matrix of lat/long coordinates
}
\description{
A wrapper around the Google and OpenSteetMpa geocoding web-services. Users can request a lat/long pair, spatail points with geocoded metadata, and/or a bounding box geometry.
A single or mulitple locations can be input. If a single point is given `geocode` with provide a matix of lat lon a spatial point an the geocode derived bounding box (if requested).
If multiple points are given the retured objects will be a matrix with columns for input name:lat:lon; a collection of spatial points; and a minimum bounding box of the input locations.
}
\examples{
\dontrun{
 #request a single location
     geocode("UCSB")
 #request a spatial point of location
     geocode("UCSB", pt = TRUE)
 #request a geocode derived bounding box of location
     geocode("UCSB", bb = TRUE)
 #request multiple locations
     geocode(c("UCSB", "Goleta", "Sterns Warf"))
 #request a minimum bounding box of all requested points
     geocode(c("UCSB", "Goleta", "Sterns Warf"), bb = T, pt= T)
}
}
\author{
Mike Johnson
}
