% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoCode.R
\name{geocode_rev}
\alias{geocode_rev}
\title{Reverse Geocoding}
\usage{
geocode_rev(xy, pt = FALSE, method = default_method)
}
\arguments{
\item{xy}{\code{logical}. If TRUE a named xy numeric vector is created}

\item{pt}{\code{logical}. If TRUE point geometry is created.}

\item{method}{the geocoding service to be used. See ?tidygeocoder::geocode}
}
\value{
a data.frame, sf object, or vector
}
\description{
Describe a location using the ERSI and OSM reverse geocoding web-services.
This service provides traditional reverse geocoding (lat/lon to placename)
but can also be use to get more information about a place name. xy must contain geographic coordinates!
}
\examples{
\dontrun{
 geocode_rev(xy = c(38,-115))
}
}
\seealso{
Other geocode: 
\code{\link{geocode_wiki}()},
\code{\link{geocode}()}
}
\concept{geocode}
