% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode_wiki.R
\name{geocode_wiki}
\alias{geocode_wiki}
\title{Geocoding Events}
\usage{
geocode_wiki(event = NULL, pt = FALSE)
}
\arguments{
\item{event}{\code{character}. a term to search for on wikipeida}

\item{pt}{\code{logical}. If TRUE point geometery is appended to the returned list()}
}
\value{
aa data.frame of lat/lon coordinates
}
\description{
A wrapper around the Wikipedia API to return geo-coordinates of requested inputs.
}
\examples{
\dontrun{
   ## geocode an Agency
     geocode_wiki("NOAA")

 ## geocode an event
     geocode_wiki("I have a dream speech")

 ## geocode a n event
     geocode_wiki("D day")

 ## geocode a product
     geocode_wiki("New York Times")

 ## geocode an event
     geocode_wiki("Hurricane Harvey")

}
}
\author{
Mike Johnson
}
