% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBoundingBox.R
\name{getBoundingBox}
\alias{getBoundingBox}
\title{Get mimimum bounding box of spatial features}
\usage{
getBoundingBox(x, sf = FALSE)
}
\arguments{
\item{x}{a \code{data.frame} with a lat and lon column, a raster, sf, or spatial object}

\item{sf}{\code{logical}. If TRUE object returned is of class \code{sf}, default is FALSE and returns \code{SpatialPolygons}
Default is \code{FALSE} and returns class SpatialPolygon}
}
\description{
Returns a minimum bounding box for a spatial, raster or sf object(s)
}
\examples{
\dontrun{
  ## Find the 10 closest Airports to UCSB
     ap = geocode("UCSB") \%>\% HydroData::findNearestAirports(n =10)
     AOI = ap$ap \%>\% getBoundingBox()

  ## Get bounding box of raster object
     AOI = getBoundingBox(r)
}
}
\author{
Mike Johnson
}
