% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBoundingBox.R
\name{getBoundingBox}
\alias{getBoundingBox}
\title{Get mimimum bounding box of spatial Objects}
\usage{
getBoundingBox(x, sf = FALSE)
}
\arguments{
\item{x}{a \code{data.frame} with a lat and long column, a Raster, sf, or Spatial Objects}

\item{sf}{\code{logical}. If \code{TRUE} object returned is of class sf
default is \code{FALSE} and returns class SpatialPolygon}
}
\description{
Returns a minimum bounding box for a set or  Spatial*, raster or sf object(s)
}
\author{
Mike Johnson
}
