% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_get_boundingbox.R
\name{getBoundingBox}
\alias{getBoundingBox}
\title{Get bounding box}
\usage{
getBoundingBox(x, sf = FALSE)
}
\arguments{
\item{x}{a \code{data.frame} with a lat and long column, a Raster or a Spatial Object}

\item{sf}{\code{logical}. If \code{TRUE} object returned is of class sf,
default is \code{FALSE} and returns class SpatialPolygonsounding box of input points \code{x}}
}
\description{
Reeurn a minimum bounding box spatial object for a set or single Spatial, raster or sf object(s)
}
\author{
Mike Johnson
}
