% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClip.R
\name{getClip}
\alias{getClip}
\title{Get a bouding box for a location}
\usage{
getClip(location = NULL, width = NULL, height = NULL, origin = NULL)
}
\arguments{
\item{location}{Defined by a location or lat, long pair}

\item{width}{define the width of the desired bounding box in miles}

\item{height}{define the height of the desired bounding box in miles}

\item{origin}{define the position of the point with respect to the bounding box. Default is set to center. Options include \itemize{
\item{"center"}
 \item{"lowerleft"}
  \item{"lowerright"}
   \item{"upperright"}
    \item{"upperleft"}
  }}
}
\value{
\code{getClip} returns a \code{SpatialPolygon} Object
}
\description{
Get a bouding box for a location
}
\details{
\code{getClip} generates a \code{SpatialPolygon} based on a location, bounding box height and width and point position.
 All HydroData outputs are projected to \emph{'+proj=longlat +ellps=GRS80 +towgs84=0,0,0,0,0,0,0+no_defs'}.
 Locations given by a character string are geocoded via the \code{dismo} package to get a lat, long pair. All bounding boxes defined by a width an a height.
 The point from chich these are drawn ins defined by a given location and origin.
}
\examples{
\dontrun{
# Get clip unit using name
    get

# Get Multi-state
    getFiatBoundary(state = c("CA","Utah","Nevada"))

# Get County
    getFiatBoundary(state = "CA", county = "San Luis Obispo")

# Get Muli-county
   getFiatBoundary(state = "CA", county = c("San Luis Obispo", "Santa Barbara", "Ventura"))

# Get counties that intersect with defined clip_unit
   getFiatBoundary(clip_unit = list("UCSB", 10, 10, "lowerleft"))
}

}
\seealso{
\itemize{
         \item \code{\link{getFiat}}
         \item \code{\link{getAOI}}
         }
}
\author{
Mike Johnson

Mike Johnson
}
