% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClip.R
\name{getClip}
\alias{getClip}
\title{Get a bouding box for a location}
\usage{
getClip(location = NULL, width = NULL, height = NULL, origin = NULL)
}
\arguments{
\item{location}{Defined by a location or lat, long pair}

\item{width}{define the width of the desired bounding box in miles}

\item{height}{define the height of the desired bounding box in miles}

\item{origin}{define the position of the point with respect to the bounding box. Default is set to center. Options include \itemize{
\item{"center"}
 \item{"lowerleft"}
  \item{"lowerright"}
   \item{"upperright"}
    \item{"upperleft"}
  }}
}
\value{
a \code{SpatialPolygons} object projected to \emph{EPSG:4269}.
}
\description{
\code{getClip} generates a \code{SpatialPolygon} based on bounding box dimisions its relation to a point.
 All HydroData outputs are projected to \emph{EPSG:4269}.
 Locations given by a character string are geocoded via the \code{dismo} package to get a lat, long pair. All bounding boxes defined by a width an a height.
 The point from chich these are drawn ins defined by a given location and origin.
}
\examples{
\dontrun{


# Get AOI defined by 10 mile bounding box using "UCSB" as the point
    getClip(location = "UCSB", width = 10, height = 10, origin = "center")

# Get AOI defined by 10 mile2 bounding box using the 'KMART near UCSB' as lower left corner
    getClip(clocation = NULL, width = NULL, height = NULL, origin = NULL)
}

}
\seealso{
\itemize{
         \item \code{\link{getClip}}
         }
}
\author{
Mike Johnson
}
