% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_getFiat.R
\name{getFiat}
\alias{getFiat}
\title{Get State of County Spatial Boundary}
\usage{
getFiat(state = NULL, county = NULL, bb = FALSE)
}
\arguments{
\item{state}{\code{character}. Full name or two character abbriviation. Not case senstive}

\item{county}{\code{character}. Provide county name(s). Requires 'state' input.}

\item{bb}{\code{logical}. If \code{TRUE} then the bounding geometry of state/county is returned,  default is \code{FALSE} and returns fiat geometries}
}
\value{
a \code{SpatialPolygons} object projected to \emph{EPSG:4269}.
}
\description{
Get State of County Spatial Boundary
}
\details{
\code{getFiat} returns a \code{SpatialPolygons} object for a defiend state and/or county.
Boundaries come from the 2017 US Census TIGER Dataset and are projected to \emph{EPSG:4269}.
}
\examples{
\dontrun{
# Get Single State
    getFiat(state = "CA")

# Get Multi-state
    getFiat(state = c("CA","Utah","Nevada"))

# Get County
    getFiat(state = "CA", county = "San Luis Obispo")

# Get Muli-county
   getFiat(state = "CA", county = c("San Luis Obispo", "Santa Barbara", "Ventura"))
}

}
\seealso{
\code{\link{getClip}}

\code{\link{getAOI}}
}
\author{
Mike Johnson
}
\keyword{internal}
