% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_nameAOI.R
\name{nameAOI}
\alias{nameAOI}
\title{Provide the plain-english description of an AOI}
\usage{
nameAOI(state = NULL, county = NULL, clip = NULL, km = FALSE)
}
\arguments{
\item{state}{\code{character}.  Full name or two character abbriviation. Not case senstive}

\item{county}{\code{character}.  County name(s). Requires \code{state} input. Not case senstive}

\item{clip}{\code{Spatial} object, a \code{Raster} object, or a \code{list} (see details and \code{\link{getClip}})}

\item{km}{\code{logical}. If \code{TRUE} distance are in kilometers,  default is \code{FALSE} and with distances in miles}
}
\value{
A character string
}
\description{
This function is a mirror of \code{\link{getAOI}} but instead of returning a Spatial object,
returns a plain-english description of the AOI. (eg: 'AOI defined as a 10 mile tall by 10 mile wide region centered on (the) 'UCSB')
}
\seealso{
\code{\link{nameClip}}
}
\author{
Mike Johnson
}
