% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_nameClip.R
\name{nameClip}
\alias{nameClip}
\title{Provide a plain-english description of a clip region}
\usage{
nameClip(clip, km = FALSE)
}
\arguments{
\item{clip}{a user supplied clip list (see \code{\link{getAOI}} and \code{\link{getClip}})}

\item{km}{\code{logical}. If \code{TRUE} distance are in kilometers,  default is \code{FALSE} and with distances in miles}
}
\value{
a string describing the clip region in plain english
}
\description{
Return a plain-english description of the clip list
(eg: 'AOI defined as a 10 mile tall by 10 mile wide region centered on (the) UCSB')
}
\author{
Mike Johnson
}
