% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revgeocode.R
\name{revgeocode}
\alias{revgeocode}
\title{Reverse Geocoding}
\usage{
revgeocode(point)
}
\arguments{
\item{point}{a point provided by lat,long or place name}
}
\value{
a data.frame of descriptive features
}
\description{
Describe a location using the ERSI and OSM reverse geocoding web-services.
}
\examples{
\dontrun{
pt1 = revgeocode("UCSB")

pt2 = revgeocode(c(38,-115))
}
}
\author{
Mike Johnson
}
