% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleCap.R, R/utility_simpleCap.R
\name{simpleCap}
\alias{simpleCap}
\alias{simpleCap}
\title{Camel Case (Simple Capitalization)}
\usage{
simpleCap(x)

simpleCap(x)
}
\arguments{
\item{x}{a text \code{character} string}

\item{x}{a text \code{character} string}
}
\description{
\code{simpleCap} capitilizes all words in a given text string

\code{simpleCap} capitilizes all words in a given text string
}
\examples{
\dontrun{
simpleCap("santa barbara")
}

\dontrun{
simpleCap("santa barbara")
}

}
\seealso{
Other AOI 'utility': \code{\link{firstLower}}
}
\author{
Mike Johnson

Mike Johnson
}
