% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_simpleCap.R
\name{simpleCap}
\alias{simpleCap}
\title{Camel Case (Simple Capitalization)}
\usage{
simpleCap(x)
}
\arguments{
\item{x}{a text \code{character} string}
}
\description{
\code{simpleCap} capitilizes all words in a given text string
}
\examples{
\dontrun{
simpleCap("santa barbara")
}

}
\author{
Mike Johnson
}
