% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{states}
\alias{states}
\title{USA States}
\format{a \code{SpatialPolygonsDataFrame}, 52 observations of 5 variables
\itemize{
\item 'statefp':    A \code{character}  State 2-digit FederalInformationProcessingStandards (FIPS) code
\item 'statens':    A \code{character}  American National Standards Institute (ANSI) code
\item 'affgeoid': A \code{character}    AFF Summary Level Code
\item 'state_name':    A \code{character}    State Name
\item 'state_abbr':     A \code{character}    State Abbriviation
}}
\source{
\href{https://cran.r-project.org/web/packages/USAboundaries/index.html}{USAboundaries}
}
\usage{
states
}
\description{
Dataset containing \code{SpatialPolygons} of USA States. Data is initalized from the USAboundaries and USAboundariesData
package, converted to spatial \code{sp} objects, and cleaned-up for this package purposes.The primary reason for doing this is to limit
the challenges associated with using the USAboundariesData (not on CRAN) as a dependency for this, and other packages, while also providing
a more minmualistic dataset.
}
\examples{
\dontrun{
 states = AOI::states
}
}
\keyword{datasets}
