% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_AOI.R
\docType{data}
\name{states}
\alias{states}
\title{USA States}
\format{a \code{SpatialPolygonsDataFrame}, 52 observations of 5 variables:
\itemize{
\item 'statefp'   :  \code{character}  2-digit Federal Information Processing Standards (FIPS) code
\item 'statens'   :  \code{character}  American National Standards Institute (ANSI) code
\item 'affgeoid'  :  \code{character}  American FactFinder (AFF) Summary Level Code
\item 'state_name':  \code{character}  State Name
\item 'state_abbr':  \code{character}  State Abbriviation
}}
\usage{
states
}
\description{
Dataset containing \code{SpatialPolygons} of USA States. Data is initalized from the USAboundaries and USAboundariesData
package, converted to \code{SpatialPolygons}, re-projected and cleaned-up for this package. The primary reason for doing this is to provide
a more minimalistic dataset primed for this package and leaflet use.
}
\examples{
\dontrun{
AOI::states
}
}
\keyword{datasets}
