% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_AOI.R
\docType{data}
\name{states}
\alias{states}
\title{USA States}
\format{a simple feature polygon set with 52 observations of 5 variables:
\itemize{
\item 'statefp'   :  \code{character}  2-digit Federal Information Processing Standards (FIPS) code
\item 'statens'   :  \code{character}  American National Standards Institute (ANSI) code
\item 'affgeoid'  :  \code{character}  American FactFinder (AFF) Summary Level Code
\item 'state_name':  \code{character}  State Name
\item 'state_abbr':  \code{character}  State Abbriviation
}}
\usage{
states
}
\description{
Dataset containing polygon representations of USA States.
}
\examples{
\dontrun{
AOI::states
}
}
\keyword{datasets}
