% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{zipcodes}
\alias{zipcodes}
\title{USA Zipcode Centroids}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 33144 rows and 7 columns.
}
\source{
\url{USABoundaries Data Package}
}
\usage{
zipcodes
}
\description{
A dataset containing the centriods of USA zipcodes
}
\keyword{datasets}
