% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_AOI.R
\docType{data}
\name{zipcodes}
\alias{zipcodes}
\title{USA Zipcode Database}
\format{a \code{data.frame}, 43191 rows of 7 variables:
\itemize{
\item 'zip'     :  \code{integer} Zipcode
\item 'city'     :  \code{character} City Name
\item 'state'     :  \code{character} State Name
\item 'latitude'       :  \code{numeric}   latitude
\item 'longitude'     :  \code{numeric} longitude
\item 'timezone'     :  \code{integer}  timezone (from UTC)
\item 'dst'  :  \code{integer}   Daylights time saving flag (1 = Observed,  0 = Not Observed)
}}
\source{
\href{https://www.boutell.com/zipcodes/}{Tom Boutell}

\href{https://www.naturalearthdata.com/downloads/50m-cultural-vectors/50m-admin-0-countries-2/}{Natural Earth}. The data is joined to a working database of the

\href{https://www.cia.gov/library/publications/the-world-factbook/}{CIA World Factbook}.
}
\usage{
zipcodes

zipcodes
}
\description{
Dataset containing information for USA Zipcodes. Data is initalized from \href{https://www.boutell.com/zipcodes/}{Tom Boutell}

Dataset containing information for USA Zipcodes. Data is initalized from \href{https://www.boutell.com/zipcodes/}{Tom Boutell}
}
\examples{
\dontrun{
 AOI::zipcodes
}
\dontrun{
 AOI::zipcodes
}
}
\keyword{datasets}
